+++ src/python/gudhi/_strong_witness_complex.cc
@@ -11,7 +11,7 @@
 #include <gudhi/Simplex_tree.h>
 #include <gudhi/Strong_witness_complex.h>
 
-#include <Simplex_tree_interface.h>
+#include <python_interfaces/Simplex_tree_interface.h>
 
 #include <utility> // std::pair
 #include <vector>
